math rendering in javascript strings
Be aware! In order to have math in the calendar and seminars box in the homepage, you need to use \\( ... \\) instead of $$...$$ and in general double all backslashes. This is because in javascript the single backslash is used for commands so this will totally fuck up the page.
The same holds for the news box on the homepage